fix: correct SDL_Keymod type from c_int to Uint16 (#4)

+++ src/core.zig
@@ -19,7 +19,7 @@ pub const SDL_SensorID = Uint32;
 pub const SDL_PenID = Uint32;
 pub const SDL_JoystickID = Uint32;
 pub const SDL_GamepadID = Uint32;
-pub const SDL_Keymod = c_int;
+pub const SDL_Keymod = Uint16;
 pub const SDL_Keycode = c_int;
 pub const SDL_Scancode = c_int;
